Change handling of the base URL
We still provide a default URL, pointing to the Test and Evaluation instance of
GRiD. If consumers of the SDK wish to point to another instance, a different
base URL can be provided as part of the GRiD client constructor NewClient().
The base URL field of the client can also be set directly at any time.

The configure subcommand of the CLI has been updated to store the base URL as
part of the configuration file. Flags have been added to configure to allow
updates to the base URL and API key fields alone (username and password coming
soon).