Move configuration back to GRiD SDK (from CLI)
Add methods to

* create/update the configuration file (in it's expected location, which is
  platform-specific),
* get the current configuration, and
* add functions to isolate reading configuration values and the masked
  password.

NewClient is now just New and returns a nil GRiD object and error if there is a
problem. This adheres to Go conventions
(https://golang.org/doc/effective_go.html#package-names).

New now retrieves configuration details from the configuration file directly.
Consumers may still check that configuration is valid prior to creating, but
they are not required to pass these directly to the constructor.

Use Mkdir rather than MkdirAll to make the .grid directory only if it doesn't
already exist (seemed to mostly be a problem on Windows).

Add hack to be able to reliably call ReadStrings on Windows immediately
following call to ReadPassword.

On Windows, configuration files are stored in USERPROFILE only, not HOMEPATH,
or as a backup, USERPROFILE.